% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_scripts.R
\name{replace_text}
\alias{replace_text}
\title{Replaces text from all \code{.R} and \code{.Rmd} files within a given
directory}
\usage{
replace_text(from, to, dir = "./")
}
\arguments{
\item{from}{one or more regexes to substitute}

\item{to}{one or values that will substitute for regexes in \code{from}}
}
\description{
This function takes a series of regular expression in \code{from} and changes
them to the values in \code{to} for all \code{.R} and \code{.Rmd} files
in the directory \code{dir}.
}
