% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scripts.R
\name{run_R}
\alias{run_R}
\title{Launches scripts from \code{.R} or \code{.Rmd} files}
\usage{
run_R(x, out_dir = "../results", output_format = "html_document", ...)
}
\arguments{
\item{x}{name of file}

\item{out_dir}{location in which to store documents compiled from \code{.Rmd}
files}

\item{output_format}{format in which to compile a \code{.Rmd} document}

\item{...}{additional parameters to pass to \code{rmarkdown::render} when
rendering a \code{.Rmd} file}
}
\value{
NULL

This function is meant to facilitate launching a series of preprocessing,
data exploration, and analysis scripts in R. After each script is run,
any existing global variables are removed; therefore, data passed from one
script to another should be outputed as an intermediate .RData file.
}
\description{
Depending on the type of file, the function will source code from a \code{.R}
file or render a document from a \code{.Rmd}.
}
