% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Writes data to an xlsx file}
\usage{
write_xlsx(datasets, file_name, sheet_names = NULL)
}
\arguments{
\item{datasets}{A list of datasets to be written to a .xlsx file}

\item{file_name}{Name and path of output file.}

\item{sheet_names}{Names of .xlsx sheets}
}
\value{
NULL

This function will create a .xlsx file with multiple sheets based on a list
of datasets specified by \code{datasets}. The names of the sheets may
be specified using \code{sheet_names}. Otherwise, the names contained in the
list of datasets will be used to name the sheets.

This function requires installlation of the \code{xlsx}.
}
\description{
Writes data to an xlsx file. If more than one dataset is specified, multiple
sheets in the output xlsx file are created.
}
